Remove unnecessary parameterization of type
The formatObject method formerly had a single object argument of generic type T (declared with the signature "<T>"). However, since T extends Object, its erasure is simply Object. More formally, "the erasure of a type variable ... is the erasure of its leftmost bound" (JLS §4.6). Therefore, the type variable in this context is redundant, as the least generic operation required in the method is a cast to Object[].