Remove unnecessary parameterization of type
The formatObject method formerly had a single argument: an object declared to be of generic type T. However, since T extends Object, its erasure is simply Object. More formally, "the erasure of a type variable...is the erasure of its leftmost bound" (JLS §4.6). Therefore, the type variable in this context is redundant.